Debugged vector database searching for similar spelling and pronounciation

+++ tests/vector_database_search.cpp
@@ -26,7 +26,7 @@ int main(int argc, char** argv) {
     const bool optional_args_provided = args.size() > 2;
 
     const int top_n = optional_args_provided ? std::stoi(args.at(2)) : 10;
-    const float threshold = optional_args_provided ? std::stoi(args.at(3)) : 0.2F;
+    const float threshold = optional_args_provided ? std::stof(args.at(3)) : 0.2F;
     const float soundex_weight = optional_args_provided ? std::stof(args.at(4)) : 0.5F;
     const float levenshtein_weight = optional_args_provided ? std::stof(args.at(5)) : 0.5F;
     const bool stop_when_top_n_are_found =
